% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brute force.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute Force Knapsack Algorithm}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{Dataframe with variables v and w.}

\item{W}{Integer as capacity.}

\item{parallel}{Boolean makes function parallel}
}
\value{
Returns a list with best value and combination.
}
\description{
Knapsack problem solution using Brute Force Algorithm in R.
            It calculates best values and weights for given capacity.
}
