% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_programming_knapsack.R
\name{dynamic_knapsack}
\alias{dynamic_knapsack}
\title{Dynamic Knapsack Algorithm}
\usage{
dynamic_knapsack(x, W)
}
\arguments{
\item{x}{Dataframe with variables v and w.}

\item{W}{Integer as capacity.}
}
\value{
Returns a list with best value and combination.
}
\description{
Knapsack problem solution using Dynamic Algorithm in R.
            It calculates best values and weights for given capacity.
}
\references{
https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem
}
