% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{greedy_knapsack}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{, is the input weight}

\item{W}{, is the input value}
}
\value{
list , returns a list that contain the maximum value and the list of elements chosen
}
\description{
greedy_knapsack function to solve the knapsack problem
}
