% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_objects.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{knapsack_objects}
\format{A data frame with 2000 rows and 2 variables:
\describe{
  \item{w}{the weight}
  \item{v}{the value}
}}
\usage{
knapsack_objects
}
\description{
A sample dataset for the knapsack problem. Code take from lab 6 of AdvR course.
}
\keyword{datasets}
